requirements.txt
packaging==23.2
Pillow==10.1.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pytesseract==0.3.10
reportlab==4.0.7
